requirements.txt
asgiref==3.2.10
astroid==2.4.2
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.4
django-allauth==0.39.1
django-cors-headers==3.6.0
django-nested-admin==3.3.2
django-rest-auth==0.9.5
django-rest-knox==4.1.0
django-rest-passwordreset==1.1.0
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
django-utils-six==2.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
gunicorn==19.7.1
idna==2.10
inflection==0.5.1
isort==5.6.4
itypes==1.2.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.0.1
openapi-codec==1.3.2
packaging==20.8
Pillow==8.0.1
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.0.0
pylint==2.6.0
PyMySQL==0.10.1
pyparsing==2.4.7
python-monkey-business==1.0.0
python3-openid==3.2.0
pytz==2020.1
requests==2.25.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.0.3
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.25.10
whitenoise==3.3.1
wrapt==1.12.1
